server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "hmeqo", email = "hmeqou@gmail.com" }]
requires-python = ">=3.13"
dependencies = [
    "setuptools",
    "pydantic-settings>=2.8.1",
    "django[argon2]>=5.1.7",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.5.0",
    "django-simpleui>=2025.1.13",
    "django-redis>=5.4.0",
    "django-cleanup>=9.0.0",
    "django-dbbackup>=4.2.1",
    "django-unused-media>=0.2.2",
    "drf-apischema>=0.2.2",
    "mysqlclient>=2.2.7",
    "whitenoise>=6.9.0",
    "granian[uvloop]>=2.1.1",
    "pillow>=11.1.0",
]

[dependency-groups]
dev = [
    "django-stubs-ext>=5.1.3",
    "django-stubs[compatible-mypy]>=5.1.3",
    "django-types>=0.20.0",
    "djangorestframework-stubs[compatible-mypy]>=3.15.3",
]

[project.scripts]
dev = "project.scripts:dev"
serve = "project.scripts:serve"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.ruff]
extend-exclude = ["migrations"]
line-length = 120
ignore = ["F401", "F403", "F405"]

[tool.ruff.format]
quote-style = "double"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "server.settings"
